requirements.txt
pydantic>=2.0.0
langchain>=0.3.0
openai>=1.0.0
tiktoken>=0.0.0
python-dotenv>=1.0.0
